+++ CHANGELOG.md
@@ -8,11 +8,13 @@
 - Set log level matplotlib fontmanager to INFO ([#90])
 - Make VCR mode configurable ([#95])
 - Add osm-timestamps to indicators + tests for it. Add timezone to oqt-timestamp([#101])
+- Add API-endpoints for list of indicators, reports, layers, datasets and fid_field ([#106])
 
 [#62]: https://github.com/GIScience/ohsome-quality-analyst/issues/62
 [#90]: https://github.com/GIScience/ohsome-quality-analyst/issues/90
 [#95]: https://github.com/GIScience/ohsome-quality-analyst/pull/95
 [#101]: https://github.com/GIScience/ohsome-quality-analyst/pull/101
+[#106]: https://github.com/GIScience/ohsome-quality-analyst/issues/106
 
 
 ## 0.4.0